JAMES-3138 InMemoryCurrentQuotaManagerTest should implement CurrentQuotaManagerContract
